requirements.txt
# Dependencias del Generador de Informes de Obras v2.0
# Instalación: pip install -r requirements.txt

# Procesamiento de datos
pandas>=1.5.0
openpyxl>=3.0.10

# Generación de PDFs
pdfkit>=1.0.0

# Templates HTML
Jinja2>=3.1.0

# Google Sheets API
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.7.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Utilidades
pathlib2>=2.3.7; python_version<"3.4"
